+++ internal/types/dislike_estimation.go
@@ -2,5 +2,5 @@ package types
 
 type DislikeEstimation struct {
 	IDHash   string `json:"id_hash"`
-	Dislikes uint32 `json:"dislikes"`
+	Dislikes int64  `json:"dislikes"`
 }
